+++ tests/pytest_pysd/pytest_cli.py
@@ -1,8 +1,10 @@
 import sys
 import os
 import shutil
-import unittest
 import subprocess
+from pathlib import Path
+
+import pytest
 
 import pandas as pd
 import numpy as np
@@ -10,7 +12,11 @@ import numpy as np
 from pysd.tools.benchmarking import load_outputs, assert_frames_close
 from pysd import __version__
 
-_root = os.path.dirname(__file__)
+# TODO replace test paths by fixtures and translate and run the models
+# in temporal directories
+# TODO make CLI tests run on Windows
+
+_root = Path(__file__).parent.parent
 
 test_model = os.path.join(_root, "test-models/samples/teacup/teacup.mdl")
 test_model_xmile = os.path.join(
@@ -54,8 +60,11 @@ def split_bash(string):
     return spl
 
 
-@unittest.skipIf(sys.platform.startswith("win"), "not working on Windows")
-class TestPySD(unittest.TestCase):
+@pytest.mark.skipif(
+    sys.platform.startswith("win"),
+    reason="not working on Windows"
+)
+class TestPySD():
     """ These tests are similar to unit_test_pysd but adapted for cli """
     def test_read_not_model(self):
 
@@ -64,11 +73,10 @@ class TestPySD(unittest.TestCase):
         command = f"{call} {model}"
         out = subprocess.run(split_bash(command), capture_output=True)
         stderr = out.stderr.decode(encoding_stderr)
-        self.assertNotEqual(out.returncode, 0)
-        self.assertIn(f"PySD: error: when parsing {model}", stderr)
-        self.assertIn(
-            "The model file name must be a Vensim (.mdl), a Xmile (.xmile, "
-            ".xml, .stmx) or a PySD (.py) model file...", stderr)
+        assert out.returncode != 0
+        assert f"PySD: error: when parsing {model}" in stderr
+        assert "The model file name must be a Vensim (.mdl), a Xmile "\
+            "(.xmile, .xml, .stmx) or a PySD (.py) model file..." in stderr
 
     def test_read_model_not_exists(self):
 
@@ -77,10 +85,9 @@ class TestPySD(unittest.TestCase):
         command = f"{call} {model}"
         out = subprocess.run(split_bash(command), capture_output=True)
         stderr = out.stderr.decode(encoding_stderr)
-        self.assertNotEqual(out.returncode, 0)
-        self.assertIn(f"PySD: error: when parsing {model}", stderr)
-        self.assertIn(
-            "The model file does not exist...", stderr)
+        assert out.returncode != 0
+        assert f"PySD: error: when parsing {model}" in stderr
+        assert "The model file does not exist..." in stderr
 
     def test_read_not_valid_output(self):
 
@@ -88,10 +95,9 @@ class TestPySD(unittest.TestCase):
         command = f"{call} -o {out_xls_file} {test_model}"
         out = subprocess.run(split_bash(command), capture_output=True)
         stderr = out.stderr.decode(encoding_stderr)
-        self.assertNotEqual(out.returncode, 0)
-        self.assertIn(f"PySD: error: when parsing {out_xls_file}", stderr)
-        self.assertIn(
-            "The output file name must be .tab or .csv...", stderr)
+        assert out.returncode != 0
+        assert f"PySD: error: when parsing {out_xls_file}" in stderr
+        assert "The output file name must be .tab or .csv..." in stderr
 
     def test_read_not_valid_time_stamps(self):
 
@@ -99,19 +105,19 @@ class TestPySD(unittest.TestCase):
         command = f"{call} -R '{time_stamps}' {test_model}"
         out = subprocess.run(split_bash(command), capture_output=True)
         stderr = out.stderr.decode(encoding_stderr)
-        self.assertNotEqual(out.returncode, 0)
-        self.assertIn(f"PySD: error: when parsing {time_stamps}", stderr)
-        self.assertIn(
-            "The return time stamps much be separated by commas...\n", stderr)
+        assert out.returncode != 0
+        assert f"PySD: error: when parsing {time_stamps}" in stderr
+        assert "The return time stamps much be separated by commas...\n"\
+            in stderr
 
         time_stamps = "1 3 4"
         command = f"{call} --return-timestamps='{time_stamps}' {test_model}"
         out = subprocess.run(split_bash(command), capture_output=True)
         stderr = out.stderr.decode(encoding_stderr)
-        self.assertNotEqual(out.returncode, 0)
-        self.assertIn(f"PySD: error: when parsing {time_stamps}", stderr)
-        self.assertIn(
-            "The return time stamps much be separated by commas...\n", stderr)
+        assert out.returncode != 0
+        assert f"PySD: error: when parsing {time_stamps}" in stderr
+        assert "The return time stamps much be separated by commas...\n"\
+            in stderr
 
     def test_read_not_valid_new_value(self):
 
@@ -119,63 +125,59 @@ class TestPySD(unittest.TestCase):
         command = f"{call} {test_model} {new_value}"
         out = subprocess.run(split_bash(command), capture_output=True)
         stderr = out.stderr.decode(encoding_stderr)
-        self.assertNotEqual(out.returncode, 0)
-        self.assertIn(f"PySD: error: when parsing {new_value}", stderr)
-        self.assertIn(
-            "You must use variable=new_value to redefine values", stderr)
+        assert out.returncode != 0
+        assert f"PySD: error: when parsing {new_value}" in stderr
+        assert "You must use variable=new_value to redefine values" in stderr
 
         new_value = "[1,2,3]"
         command = f"{call} {test_model} {new_value}"
         out = subprocess.run(split_bash(command), capture_output=True)
         stderr = out.stderr.decode(encoding_stderr)
-        self.assertNotEqual(out.returncode, 0)
-        self.assertIn(f"PySD: error: when parsing {new_value}", stderr)
-        self.assertIn(
-            "You must use variable=new_value to redefine values", stderr)
+        assert out.returncode != 0
+        assert f"PySD: error: when parsing {new_value}" in stderr
+        assert "You must use variable=new_value to redefine values" in stderr
 
         new_value = "foo:[[1,2,3],[4,5,6]]"
         command = f"{call} {test_model} {new_value}"
         out = subprocess.run(split_bash(command), capture_output=True)
         stderr = out.stderr.decode(encoding_stderr)
-        self.assertNotEqual(out.returncode, 0)
-        self.assertIn(f"PySD: error: when parsing {new_value}", stderr)
-        self.assertIn(
-            "You must use variable=new_value to redefine values", stderr)
+        assert out.returncode != 0
+        assert f"PySD: error: when parsing {new_value}" in stderr
+        assert "You must use variable=new_value to redefine values" in stderr
 
     def test_print_version(self):
 
         command = f"{call} -v"
         out = subprocess.run(split_bash(command), capture_output=True)
         stdout = out.stdout.decode(encoding_stdout)
-        self.assertEqual(out.returncode, 0)
-        self.assertIn(f"PySD {__version__}", stdout)
+        assert out.returncode == 0
+        assert f"PySD {__version__}" in stdout
 
         command = f"{call} --version"
         out = subprocess.run(split_bash(command), capture_output=True)
         stdout = out.stdout.decode(encoding_stdout)
-        self.assertEqual(out.returncode, 0)
-        self.assertIn(f"PySD {__version__}", stdout)
+        assert out.returncode == 0
+        assert f"PySD {__version__}" in stdout
 
     def test_print_help(self):
 
         command = f"{call}"
         out = subprocess.run(split_bash(command), capture_output=True)
         stdout = out.stdout.decode(encoding_stdout)
-        self.assertEqual(out.returncode, 0)
-        self.assertIn("usage: python -m pysd [", stdout)
+        assert out.returncode == 0
+        assert "usage: python -m pysd [" in stdout
 
         command = f"{call} -h"
         out = subprocess.run(split_bash(command), capture_output=True)
         stdout = out.stdout.decode(encoding_stdout)
-        self.assertEqual(out.returncode, 0)
-        self.assertIn("usage: python -m pysd [", stdout)
+        assert out.returncode == 0
+        assert "usage: python -m pysd [" in stdout
 
         command = f"{call} --help"
         out = subprocess.run(split_bash(command), capture_output=True)
         stdout = out.stdout.decode(encoding_stdout)
-        self.assertEqual(out.returncode, 0)
-        self.assertIn(
-            "usage: python -m pysd [", stdout)
+        assert out.returncode == 0
+        assert "usage: python -m pysd [" in stdout
 
     def test_translate_file(self):
 
@@ -188,9 +190,9 @@ class TestPySD(unittest.TestCase):
 
         command = f"{call} --translate {test_model}"
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
-        self.assertFalse(os.path.isfile(out_tab_file))
-        self.assertTrue(os.path.isfile(model_py))
+        assert out.returncode == 0
+        assert not os.path.isfile(out_tab_file)
+        assert os.path.isfile(model_py)
         os.remove(model_py)
 
     def test_read_vensim_split_model(self):
@@ -205,27 +207,23 @@ class TestPySD(unittest.TestCase):
 
         command = f"{call} --translate --split-views {model_name_mdl}"
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
 
         # check that _subscript_dict json file was created
-        self.assertTrue(os.path.isfile(root_dir + subscript_filename))
+        assert os.path.isfile(root_dir + subscript_filename)
 
         # check that the main model file was created
-        self.assertTrue(os.path.isfile(root_dir + model_name + ".py"))
+        assert os.path.isfile(root_dir + model_name + ".py")
 
         # check that the modules folder was created
-        self.assertTrue(os.path.isdir(root_dir + modules_dirname))
-        self.assertTrue(
-            os.path.isfile(root_dir + modules_dirname + "/" + modules_filename)
-        )
+        assert os.path.isdir(root_dir + modules_dirname)
+        assert os.path.isfile(root_dir + modules_dirname
+                              + "/" + modules_filename)
 
         # check creation of module files
-        self.assertTrue(
-            os.path.isfile(root_dir + modules_dirname + "/" + "view_1.py"))
-        self.assertTrue(
-            os.path.isfile(root_dir + modules_dirname + "/" + "view2.py"))
-        self.assertTrue(
-            os.path.isfile(root_dir + modules_dirname + "/" + "view_3.py"))
+        assert os.path.isfile(root_dir + modules_dirname + "/" + "view_1.py")
+        assert os.path.isfile(root_dir + modules_dirname + "/" + "view2.py")
+        assert os.path.isfile(root_dir + modules_dirname + "/" + "view_3.py")
 
         # remove newly created files
         os.remove(root_dir + model_name + ".py")
@@ -255,20 +253,17 @@ class TestPySD(unittest.TestCase):
         command = f"{call} --translate --split-views "\
                   f"--subview-sep={separator} {model_name_mdl}"
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
 
         # check that the modules folders were created
-        self.assertTrue(os.path.isdir(root_dir + modules_dirname + "/view_1"))
+        assert os.path.isdir(root_dir + modules_dirname + "/view_1")
 
         # check creation of module files
-        self.assertTrue(
-            os.path.isfile(root_dir + modules_dirname + "/view_1/" +
-                           "submodule_1.py"))
-        self.assertTrue(
-            os.path.isfile(root_dir + modules_dirname + "/view_1/" +
-                           "submodule_2.py"))
-        self.assertTrue(
-            os.path.isfile(root_dir + modules_dirname + "/view_2.py"))
+        assert os.path.isfile(root_dir + modules_dirname + "/view_1/" +
+                              "submodule_1.py")
+        assert os.path.isfile(root_dir + modules_dirname + "/view_1/" +
+                              "submodule_2.py")
+        assert os.path.isfile(root_dir + modules_dirname + "/view_2.py")
 
         # check that the results of the split model are the same than those
         # without splitting
@@ -297,16 +292,16 @@ class TestPySD(unittest.TestCase):
         command = f"{call} -o {out_csv_file} -R {','.join(timestamps)} "\
                   f" {test_model}"
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
         stocks = load_outputs(out_csv_file)
-        self.assertTrue((stocks.index.values.astype(str) == timestamps).all())
+        assert (stocks.index.values.astype(str) == timestamps).all()
         os.remove(out_csv_file)
 
         command = f"{call} -o {out_csv_file} -R 5 {test_model}"
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
         stocks = load_outputs(out_csv_file)
-        self.assertTrue((stocks.index.values == [5]))
+        assert (stocks.index.values == [5])
         os.remove(out_csv_file)
 
     def test_run_return_columns(self):
@@ -316,9 +311,9 @@ class TestPySD(unittest.TestCase):
                   f" {test_model}"
 
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
         stocks = load_outputs(out_csv_file)
-        self.assertEqual(set(stocks.columns), set(return_columns))
+        assert set(stocks.columns) == set(return_columns)
         os.remove(out_csv_file)
 
         # from txt
@@ -330,9 +325,9 @@ class TestPySD(unittest.TestCase):
         command = f"{call} -o {out_csv_file} -r {txt_file} {test_model}"
 
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
         stocks = load_outputs(out_csv_file)
-        self.assertEqual(set(stocks.columns), set(return_columns))
+        assert set(stocks.columns) == set(return_columns)
 
         os.remove(txt_file)
         os.remove(out_csv_file)
@@ -343,9 +338,9 @@ class TestPySD(unittest.TestCase):
                   f" {test_model}"
 
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
         stocks = load_outputs(out_csv_file)
-        self.assertEqual(set(stocks.columns), set(return_columns))
+        assert set(stocks.columns) == set(return_columns)
         os.remove(out_csv_file)
 
     def test_model_arguments(self):
@@ -353,29 +348,29 @@ class TestPySD(unittest.TestCase):
         initial_time = 10
         command = f"{call} -o {out_tab_file} -I {initial_time} {test_model}"
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
         stocks = load_outputs(out_tab_file)
-        self.assertTrue(stocks.index.values[0] == initial_time)
+        assert stocks.index.values[0] == initial_time
         os.remove(out_tab_file)
 
         # check final time
         final_time = 20
         command = f"{call} -o {out_tab_file} -F {final_time} {test_model}"
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
         stocks = load_outputs(out_tab_file)
-        self.assertTrue(stocks.index.values[-1] == final_time)
+        assert stocks.index.values[-1] == final_time
         os.remove(out_tab_file)
 
         # check time step
         time_step = 10
         command = f"{call} -o {out_tab_file} -T {time_step} {test_model}"
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
         stocks = load_outputs(out_tab_file)
-        self.assertTrue((np.diff(stocks.index.values) == time_step).all())
-        self.assertTrue((stocks["SAVEPER"] == time_step).all().all())
-        self.assertTrue((stocks["TIME STEP"] == time_step).all().all())
+        assert (np.diff(stocks.index.values) == time_step).all()
+        assert (stocks["SAVEPER"] == time_step).all().all()
+        assert (stocks["TIME STEP"] == time_step).all().all()
         os.remove(out_tab_file)
 
         # check saveper
@@ -384,11 +379,11 @@ class TestPySD(unittest.TestCase):
         command = f"{call} -o {out_tab_file} -T {time_step} "\
                   f"-S {saveper} {test_model}"
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
         stocks = load_outputs(out_tab_file)
-        self.assertTrue((np.diff(stocks.index.values) == saveper).all())
-        self.assertTrue((stocks["SAVEPER"] == saveper).all().all())
-        self.assertTrue((stocks["TIME STEP"] == time_step).all().all())
+        assert (np.diff(stocks.index.values) == saveper).all()
+        assert (stocks["SAVEPER"] == saveper).all().all()
+        assert (stocks["TIME STEP"] == time_step).all().all()
         os.remove(out_tab_file)
 
         # check all
@@ -400,11 +395,11 @@ class TestPySD(unittest.TestCase):
                   f"--saveper={saveper} --initial-time={initial_time} "\
                   f"--final-time={final_time} {test_model}"
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
         stocks = load_outputs(out_tab_file)
-        self.assertTrue((np.diff(stocks.index.values) == saveper).all())
-        self.assertTrue(stocks.index.values[0] == initial_time)
-        self.assertTrue(stocks.index.values[-1] == final_time)
+        assert (np.diff(stocks.index.values) == saveper).all()
+        assert stocks.index.values[0] == initial_time
+        assert stocks.index.values[-1] == final_time
         os.remove(out_tab_file)
 
     def test_initial_conditions_tuple_pysafe_names(self):
@@ -422,7 +417,7 @@ class TestPySD(unittest.TestCase):
                   f" teacup_temperature:33"
 
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
         stocks2 = load_outputs(out_tab_file)
         assert_frames_close(stocks2, stocks)
         os.remove(out_tab_file)
@@ -435,9 +430,9 @@ class TestPySD(unittest.TestCase):
                   f" room_temperature={value}"
 
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
         stocks = load_outputs(out_tab_file)
-        self.assertTrue((stocks["room_temperature"] == value).all())
+        assert (stocks["room_temperature"] == value).all()
         os.remove(out_tab_file)
 
     def test_set_timeseries_parameter_lookup(self):
@@ -457,9 +452,9 @@ class TestPySD(unittest.TestCase):
                   f" {test_model_look}"\
                   f" lookup_1d_time={timeseries_bash}"
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
         stocks = load_outputs(out_tab_file)
-        self.assertTrue((stocks["lookup_1d_time"] == temp_timeseries).all())
+        assert (stocks["lookup_1d_time"] == temp_timeseries).all()
         os.remove(out_tab_file)
 
         command = f"{call} -o {out_tab_file} -r lookup_2d_time "\
@@ -467,12 +462,10 @@ class TestPySD(unittest.TestCase):
                   f" {test_model_look}"\
                   f" lookup_2d_time={timeseries_bash}"
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
         stocks = load_outputs(out_tab_file)
-        self.assertTrue(
-            (stocks["lookup_2d_time[Row1]"] == temp_timeseries).all())
-        self.assertTrue(
-            (stocks["lookup_2d_time[Row2]"] == temp_timeseries).all())
+        assert (stocks["lookup_2d_time[Row1]"] == temp_timeseries).all()
+        assert (stocks["lookup_2d_time[Row2]"] == temp_timeseries).all()
         os.remove(out_tab_file)
 
     def test_export_import(self):
@@ -482,7 +475,7 @@ class TestPySD(unittest.TestCase):
         exp_file = "teacup15.pic"
         model = pysd.read_vensim(test_model)
         stocks = model.run(return_timestamps=[0, 10, 20, 30])
-        self.assertTrue((stocks["INITIAL TIME"] == 0).all().all())
+        assert (stocks["INITIAL TIME"] == 0).all().all()
 
         command = f"{call} -o {out_tab_file} -e {exp_file} -F 15 -R 0,10"\
                   f" {test_model}"
@@ -491,19 +484,19 @@ class TestPySD(unittest.TestCase):
                    f" {test_model}"
 
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
         stocks1 = load_outputs(out_tab_file)
 
         out = subprocess.run(split_bash(command2), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
         stocks2 = load_outputs(out_tab_file)
 
         os.remove(exp_file)
         os.remove(out_tab_file)
 
-        self.assertTrue((stocks1["INITIAL TIME"] == 0).all().all())
-        self.assertTrue((stocks1["FINAL TIME"] == 15).all().all())
-        self.assertTrue((stocks2["INITIAL TIME"] == 15).all().all())
+        assert (stocks1["INITIAL TIME"] == 0).all().all()
+        assert (stocks1["FINAL TIME"] == 15).all().all()
+        assert (stocks2["INITIAL TIME"] == 15).all().all()
         stocks.drop("INITIAL TIME", axis=1, inplace=True)
         stocks1.drop("INITIAL TIME", axis=1, inplace=True)
         stocks2.drop("INITIAL TIME", axis=1, inplace=True)
@@ -526,7 +519,7 @@ class TestPySD(unittest.TestCase):
                   f" {model_file}"
 
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
         stocks = load_outputs(out_tab_file)
         canon = load_outputs(os.path.join(
             _root,
@@ -539,22 +532,20 @@ class TestPySD(unittest.TestCase):
                   f" {model_file}"
 
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertNotEqual(out.returncode, 0)
+        assert out.returncode != 0
         stderr = out.stderr.decode(encoding_stderr)
-        self.assertIn("PySD: error: when parsing my_file.txt", stderr)
-        self.assertIn(
-            "The data file name must be .tab or .csv...", stderr)
+        assert "PySD: error: when parsing my_file.txt" in stderr
+        assert "The data file name must be .tab or .csv..." in stderr
 
         # not found data file
         command = f"{call} -o {out_tab_file} -D my_file.tab"\
                   f" {model_file}"
 
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertNotEqual(out.returncode, 0)
+        assert out.returncode != 0
         stderr = out.stderr.decode(encoding_stderr)
-        self.assertIn("PySD: error: when parsing my_file.tab", stderr)
-        self.assertIn(
-            "The data file does not exist...", stderr)
+        assert "PySD: error: when parsing my_file.tab" in stderr
+        assert "The data file does not exist..." in stderr
 
     def test_save_without_name(self):
         import re
@@ -562,14 +553,14 @@ class TestPySD(unittest.TestCase):
         command = f"{call} {test_model}"
         command2 = f"{call} -o {out_tab_file} {test_model}"
         out = subprocess.run(split_bash(command), capture_output=True)
-        self.assertEqual(out.returncode, 0)
+        assert out.returncode == 0
         stdout = out.stdout.decode(encoding_stdout)
         outputs = re.findall("(?<=Data saved in ').*(?=')", stdout)[0]
         out2 = subprocess.run(split_bash(command2), capture_output=True)
-        self.assertEqual(out2.returncode, 0)
+        assert out2.returncode == 0
 
         out, out2 = load_outputs(outputs), load_outputs(out_tab_file)
         os.remove(outputs)
         os.remove(out_tab_file)
 
-        self.assertTrue((out - out2 == 0).all().all())
+        assert (out - out2 == 0).all().all()