+++ tests/pytest_pysd/pytest_utils.py
@@ -1,20 +1,18 @@
 import doctest
-from pathlib import Path
-from unittest import TestCase
 
+import pytest
 import pandas as pd
 import xarray as xr
 
 from pysd.tools.benchmarking import assert_frames_close
+import pysd
+from pysd.py_backend.utils import compute_shape, rearrange, load_outputs,\
+    ProgressBar
 
-_root = Path(__file__).parent
 
-
-class TestUtils(TestCase):
+class TestUtils():
 
     def test_xrsplit(self):
-        import pysd
-
         array1d = xr.DataArray([0.5, 0., 1.],
                                {'ABC': ['A', 'B', 'C']},
                                ['ABC'])
@@ -36,85 +34,72 @@ class TestUtils(TestCase):
         s3d = pysd.utils.xrsplit(array3d)
 
         # check length
-        self.assertEqual(len(s1d), 3)
-        self.assertEqual(len(s2d), 6)
-        self.assertEqual(len(s3d), 12)
+        assert len(s1d) == 3
+        assert len(s2d) == 6
+        assert len(s3d) == 12
 
         # check all values for 1d
-        self.assertIn(xr.DataArray(0.5, {'ABC': ['A']}, ['ABC']), s1d)
-        self.assertIn(xr.DataArray(0., {'ABC': ['B']}, ['ABC']), s1d)
-        self.assertIn(xr.DataArray(1., {'ABC': ['C']}, ['ABC']), s1d)
+        assert xr.DataArray(0.5, {'ABC': ['A']}, ['ABC']) in s1d
+        assert xr.DataArray(0., {'ABC': ['B']}, ['ABC']) in s1d
+        assert xr.DataArray(1., {'ABC': ['C']}, ['ABC']) in s1d
         # check some values for 2d and 3d
-        self.assertIn(xr.DataArray(0.5,
-                                   {'ABC': ['A'], 'XY': ['X']},
-                                   ['ABC', 'XY']),
-                      s2d)
-        self.assertIn(xr.DataArray(-0.5,
-                                   {'ABC': ['B'], 'XY': ['Y']},
-                                   ['ABC', 'XY']),
-                      s2d)
-        self.assertIn(xr.DataArray(-0.5,
-                                   {'ABC': ['B'], 'XY': ['Y'], 'FG': ['F']},
-                                   ['ABC', 'XY', 'FG']),
-                      s3d)
-        self.assertIn(xr.DataArray(0.75,
-                                   {'ABC': ['C'], 'XY': ['X'], 'FG': ['G']},
-                                   ['ABC', 'XY', 'FG']),
-                      s3d)
+        assert xr.DataArray(
+            0.5, {'ABC': ['A'], 'XY': ['X']}, ['ABC', 'XY']
+            ) in s2d
+        assert xr.DataArray(
+            -0.5, {'ABC': ['B'], 'XY': ['Y']}, ['ABC', 'XY']
+            ) in s2d
+        assert xr.DataArray(
+            -0.5, {'ABC': ['B'], 'XY': ['Y'], 'FG': ['F']}, ['ABC', 'XY', 'FG']
+            ) in s3d
+        assert xr.DataArray(
+            0.75, {'ABC': ['C'], 'XY': ['X'], 'FG': ['G']}, ['ABC', 'XY', 'FG']
+            ) in s3d
 
     def test_get_return_elements_subscirpts(self):
-        import pysd
-
-        self.assertEqual(
-            pysd.utils.get_return_elements(
+        assert pysd.utils.get_return_elements(
                 ["Inflow A[Entry 1,Column 1]",
                  "Inflow A[Entry 1,Column 2]"],
-                {'Inflow A': 'inflow_a'}),
-            (['inflow_a'],
-             {'Inflow A[Entry 1,Column 1]': ('inflow_a',
-                                             ('Entry 1', 'Column 1')),
-              'Inflow A[Entry 1,Column 2]': ('inflow_a',
-                                             ('Entry 1', 'Column 2'))}
-             )
-        )
+                {'Inflow A': 'inflow_a'})\
+            == (
+            ['inflow_a'],
+            {'Inflow A[Entry 1,Column 1]': ('inflow_a',
+                                            ('Entry 1', 'Column 1')),
+             'Inflow A[Entry 1,Column 2]': ('inflow_a',
+                                            ('Entry 1', 'Column 2'))}
+            )
 
     def test_get_return_elements_realnames(self):
-        import pysd
-        self.assertEqual(
-            pysd.utils.get_return_elements(
+        assert pysd.utils.get_return_elements(
                 ["Inflow A", "Inflow B"],
-                {'Inflow A': 'inflow_a', 'Inflow B': 'inflow_b'}),
-            (['inflow_a', 'inflow_b'],
-             {'Inflow A': ('inflow_a', None),
-              'Inflow B': ('inflow_b', None)}
-             )
-        )
+                {'Inflow A': 'inflow_a', 'Inflow B': 'inflow_b'})\
+            == (
+            ['inflow_a', 'inflow_b'],
+            {'Inflow A': ('inflow_a', None),
+             'Inflow B': ('inflow_b', None)}
+            )
 
     def test_get_return_elements_pysafe_names(self):
-        import pysd
-        self.assertEqual(
-            pysd.utils.get_return_elements(
+        assert pysd.utils.get_return_elements(
                 ["inflow_a", "inflow_b"],
-                {'Inflow A': 'inflow_a', 'Inflow B': 'inflow_b'}),
-            (['inflow_a', 'inflow_b'],
-             {'inflow_a': ('inflow_a', None),
-              'inflow_b': ('inflow_b', None)}
-             )
-        )
+                {'Inflow A': 'inflow_a', 'Inflow B': 'inflow_b'})\
+            == (
+            ['inflow_a', 'inflow_b'],
+            {'inflow_a': ('inflow_a', None),
+             'inflow_b': ('inflow_b', None)}
+            )
 
     def test_get_return_elements_not_found_error(self):
         """"
         Test for not found element
         """
-        import pysd
 
-        with self.assertRaises(KeyError):
+        with pytest.raises(KeyError):
             pysd.utils.get_return_elements(
                 ["inflow_a", "inflow_b", "inflow_c"],
                 {'Inflow A': 'inflow_a', 'Inflow B': 'inflow_b'})
 
     def test_make_flat_df(self):
-        import pysd
 
         df = pd.DataFrame(index=[1], columns=['elem1'])
         df.at[1] = [xr.DataArray([[1, 2, 3], [4, 5, 6], [7, 8, 9]],
@@ -130,11 +115,10 @@ class TestUtils(TestCase):
         actual = pysd.utils.make_flat_df(df, return_addresses)
 
         # check all columns are in the DataFrame
-        self.assertEqual(set(actual.columns), set(expected.columns))
+        assert set(actual.columns) == set(expected.columns)
         assert_frames_close(actual, expected, rtol=1e-8, atol=1e-8)
 
     def test_make_flat_df_0dxarray(self):
-        import pysd
 
         df = pd.DataFrame(index=[1], columns=['elem1'])
         df.at[1] = [xr.DataArray(5)]
@@ -146,11 +130,10 @@ class TestUtils(TestCase):
         actual = pysd.utils.make_flat_df(df, return_addresses, flatten=True)
 
         # check all columns are in the DataFrame
-        self.assertEqual(set(actual.columns), set(expected.columns))
+        assert set(actual.columns) == set(expected.columns)
         assert_frames_close(actual, expected, rtol=1e-8, atol=1e-8)
 
     def test_make_flat_df_nosubs(self):
-        import pysd
 
         df = pd.DataFrame(index=[1], columns=['elem1', 'elem2'])
         df.at[1] = [25, 13]
@@ -164,15 +147,14 @@ class TestUtils(TestCase):
         actual = pysd.utils.make_flat_df(df, return_addresses)
 
         # check all columns are in the DataFrame
-        self.assertEqual(set(actual.columns), set(expected.columns))
-        self.assertTrue(all(actual['Elem1'] == expected['Elem1']))
-        self.assertTrue(all(actual['Elem2'] == expected['Elem2']))
+        assert set(actual.columns) == set(expected.columns)
+        assert all(actual['Elem1'] == expected['Elem1'])
+        assert all(actual['Elem2'] == expected['Elem2'])
 
     def test_make_flat_df_return_array(self):
         """ There could be cases where we want to
         return a whole section of an array - ie, by passing in only part of
         the simulation dictionary. in this case, we can't force to float..."""
-        import pysd
 
         df = pd.DataFrame(index=[1], columns=['elem1', 'elem2'])
         df.at[1] = [xr.DataArray([[1, 2, 3], [4, 5, 6], [7, 8, 9]],
@@ -202,16 +184,13 @@ class TestUtils(TestCase):
         actual = pysd.utils.make_flat_df(df, return_addresses)
 
         # check all columns are in the DataFrame
-        self.assertEqual(set(actual.columns), set(expected.columns))
+        assert set(actual.columns) == set(expected.columns)
         # need to assert one by one as they are xarrays
-        self.assertTrue(
-            actual.loc[1, 'Elem1[A, Dim2]'].equals(
-                expected.loc[1, 'Elem1[A, Dim2]']))
-        self.assertTrue(
-            actual.loc[1, 'Elem2'].equals(expected.loc[1, 'Elem2']))
+        assert actual.loc[1, 'Elem1[A, Dim2]'].equals(
+                expected.loc[1, 'Elem1[A, Dim2]'])
+        assert actual.loc[1, 'Elem2'].equals(expected.loc[1, 'Elem2'])
 
     def test_make_flat_df_flatten(self):
-        import pysd
 
         df = pd.DataFrame(index=[1], columns=['elem1', 'elem2'])
         df.at[1] = [xr.DataArray([[1, 2, 3], [4, 5, 6], [7, 8, 9]],
@@ -247,15 +226,12 @@ class TestUtils(TestCase):
         actual = pysd.utils.make_flat_df(df, return_addresses, flatten=True)
 
         # check all columns are in the DataFrame
-        self.assertEqual(set(actual.columns), set(expected.columns))
+        assert set(actual.columns) == set(expected.columns)
         # need to assert one by one as they are xarrays
         for col in set(expected.columns):
-            self.assertEqual(
-                actual.loc[:, col].values,
-                expected.loc[:, col].values)
+            assert actual.loc[:, col].values == expected.loc[:, col].values
 
     def test_make_flat_df_flatten_transposed(self):
-        import pysd
 
         df = pd.DataFrame(index=[1], columns=['elem2'])
         df.at[1] = [
@@ -285,15 +261,12 @@ class TestUtils(TestCase):
         actual = pysd.utils.make_flat_df(df, return_addresses, flatten=True)
         print(actual.columns)
         # check all columns are in the DataFrame
-        self.assertEqual(set(actual.columns), set(expected.columns))
+        assert set(actual.columns) == set(expected.columns)
         # need to assert one by one as they are xarrays
         for col in set(expected.columns):
-            self.assertEqual(
-                actual.loc[:, col].values,
-                expected.loc[:, col].values)
+            assert actual.loc[:, col].values == expected.loc[:, col].values
 
     def test_make_flat_df_times(self):
-        import pysd
 
         df = pd.DataFrame(index=[1, 2], columns=['elem1'])
         df['elem1'] = [xr.DataArray([[1, 2, 3], [4, 5, 6], [7, 8, 9]],
@@ -313,12 +286,11 @@ class TestUtils(TestCase):
         actual = pysd.utils.make_flat_df(df, return_addresses)
 
         # check all columns are in the DataFrame
-        self.assertEqual(set(actual.columns), set(expected.columns))
-        self.assertEqual(set(actual.index), set(expected.index))
-        self.assertTrue(all(actual['Elem1[B,F]'] == expected['Elem1[B,F]']))
+        assert set(actual.columns) == set(expected.columns)
+        assert set(actual.index) == set(expected.index)
+        assert all(actual['Elem1[B,F]'] == expected['Elem1[B,F]'])
 
     def test_doctests(self):
-        import pysd
         doctest.DocTestSuite(pysd.utils)
 
     def test_compute_shape(self):
@@ -326,8 +298,6 @@ class TestUtils(TestCase):
         Test for computing the shape of an array giving coordinates dictionary
         and ordered dimensions.
         """
-        from pysd.py_backend.utils import compute_shape
-
         coords = [
           {},
           {'ABC': ['A', 'B', 'C'],
@@ -347,15 +317,13 @@ class TestUtils(TestCase):
         shapes = [[], [3, 1, 8], [5, 3, 1], [3, 8, 2], [8, 3], [8]]
 
         for c, s in zip(coords, shapes):
-            self.assertEqual(compute_shape(c), s)
+            assert compute_shape(c) == s
 
     def test_compute_shape_reshape(self):
         """"
         Test for computing the shape of an array giving coordinates dictionary
         and ordered dimensions with reshape.
         """
-        from pysd.py_backend.utils import compute_shape
-
         coords = [
           {'ABC': ['A', 'B', 'C'],
            'XY': ['X'],
@@ -381,14 +349,12 @@ class TestUtils(TestCase):
         for i, shapes in enumerate(shapes123):
             for c, s in zip(coords, shapes):
                 if s:
-                    self.assertEqual(compute_shape(c, i+1), s)
+                    assert compute_shape(c, i+1) == s
                 else:
-                    with self.assertRaises(ValueError):
+                    with pytest.raises(ValueError):
                         compute_shape(c, i+1)
 
     def test_rearrange(self):
-        from pysd.py_backend.utils import rearrange
-
         # simple cases are tested, complex cases are tested with test-models
         _subscript_dict = {
             'd1': ['a', 'b', 'c'],
@@ -425,48 +391,40 @@ class TestUtils(TestCase):
         xr_out_float = xr.DataArray(3.,
                                     {'d2': ['b', 'c']}, ['d2'])
 
-        self.assertTrue(xr_out_subdim.equals(
-            rearrange(xr_input_subdim, ['d2'], _subscript_dict)))
+        assert xr_out_subdim.equals(
+            rearrange(xr_input_subdim, ['d2'], _subscript_dict))
 
-        self.assertTrue(xr_out_subdim2.equals(
-            rearrange(xr_input_subdim2, ['d4', 'd2'], _subscript_dict)))
+        assert xr_out_subdim2.equals(
+            rearrange(xr_input_subdim2, ['d4', 'd2'], _subscript_dict))
 
-        self.assertTrue(xr_out_updim.equals(
-            rearrange(xr_input_updim, ['d2', 'd3'], _subscript_dict)))
+        assert xr_out_updim.equals(
+            rearrange(xr_input_updim, ['d2', 'd3'], _subscript_dict))
 
-        self.assertTrue(xr_out_switch.equals(
-            rearrange(xr_input_switch, ['d3', 'd2'], _subscript_dict)))
+        assert xr_out_switch.equals(
+            rearrange(xr_input_switch, ['d3', 'd2'], _subscript_dict))
 
-        self.assertTrue(xr_out_float.equals(
-            rearrange(xr_input_float, ['d2'], _subscript_dict)))
+        assert xr_out_float.equals(
+            rearrange(xr_input_float, ['d2'], _subscript_dict))
 
-        self.assertEqual(None,
-                         rearrange(None, ['d2'], _subscript_dict))
+        assert rearrange(None, ['d2'], _subscript_dict) is None
 
 
-class TestLoadOutputs(TestCase):
-    def test_non_valid_outputs(self):
-        from pysd.py_backend.utils import load_outputs
-
+class TestLoadOutputs():
+    def test_non_valid_outputs(self, _root):
         outputs = _root.joinpath("more-tests/not_vensim/test_not_vensim.txt")
 
-        with self.assertRaises(ValueError) as err:
+        error_message = "Not able to read '%s'." % outputs
+        with pytest.raises(ValueError, match=error_message):
             load_outputs(outputs)
 
-        self.assertIn("Not able to read '%s'." % outputs, str(err.exception))
-
-    def test_transposed_frame(self):
-        from pysd.py_backend.utils import load_outputs
-
+    def test_transposed_frame(self, _root):
         assert_frames_close(
             load_outputs(_root.joinpath("data/out_teacup.csv")),
             load_outputs(
                 _root.joinpath("data/out_teacup_transposed.csv"),
                 transpose=True))
 
-    def test_load_columns(self):
-        from pysd.py_backend.utils import load_outputs
-
+    def test_load_columns(self, _root):
         out0 = load_outputs(_root.joinpath("data/out_teacup.csv"))
 
         out1 = load_outputs(
@@ -478,31 +436,27 @@ class TestLoadOutputs(TestCase):
             transpose=True,
             columns=["Heat Loss to Room"])
 
-        self.assertEqual(
-            set(out1.columns),
-            set(["Room Temperature", "Teacup Temperature"]))
+        assert set(out1.columns)\
+            == set(["Room Temperature", "Teacup Temperature"])
 
-        self.assertEqual(
-            set(out2.columns),
-            set(["Heat Loss to Room"]))
+        assert set(out2.columns)\
+            == set(["Heat Loss to Room"])
 
-        self.assertTrue((out0.index == out1.index).all())
-        self.assertTrue((out0.index == out2.index).all())
+        assert (out0.index == out1.index).all()
+        assert (out0.index == out2.index).all()
 
 
-class TestProgressbar(TestCase):
+class TestProgressbar():
     def test_progressbar(self):
-        from pysd.py_backend.utils import ProgressBar
-
         pbar = ProgressBar(10)
 
         for i in range(10):
-            self.assertEqual(pbar.counter, i)
+            assert pbar.counter == i
             pbar.update()
 
         pbar.finish()
 
         pbar = ProgressBar()
-        self.assertFalse(hasattr(pbar, 'counter'))
+        assert not hasattr(pbar, 'counter')
         pbar.update()
         pbar.finish()